Add option to filter BLE scan operation for devices with supported services
Introduces an optional parameter
startScanning(filterUnsupportedPeripherials: Bool = true)

If set to true it will pass the list of known serial services to
centralManager.scanForPeripherals(...) so that only devices with know services
are detected.

This removes the burden from the application developer to manually connect to
detected devices and check if they  are supported and allows to easily scan
for only supported devices.

By default the option is set to true since there seems to be no downside to this,
but it could easily be changed to default to false in case there are any concerns
about backwards compatibility.