Add support for a new Sirius profile version
After upgrading a Mares Sirius to firmware version 1.6.16, the profile

version changed to v2.0. Surprisingly, despite the change in the major

version number, there appear to be no functional changes in the data

format.

Add support for a new Sirius header version

The Mares Sirius firmare upgrade to v1.6.16 also changed the header

version to v2.0. Again, there appear to be no functional changes in the

data format.